services/github-client/requirements.txt

fastapi==0.115.6
uvicorn==0.34.0
PyGithub==2.5.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
prometheus-client==0.21.1

